+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 
 ## [Unreleased] - ReleaseDate
 
-[Commits](https://github.com/twitch-rs/twitch_types/compare/v0.3.5...Unreleased)
+[Commits](https://github.com/twitch-rs/twitch_types/compare/v0.3.6...Unreleased)
+
+## [v0.3.6] - 2022-10-28
+
+[Commits](https://github.com/twitch-rs/twitch_types/compare/v0.3.5...v0.3.6)
 
 ### Added